Add 'modsPath' optional setting
The default behavior looks for mod folders inside the ArmA3 folder path.
With this commit I add an optional setting inside 'config.js', named
.modsPath, which allows you to set a different path to your mods folder.
If the setting is not set inside the 'config.js' file, .modsPath value
will default to the value of .path (the path to your ArmA3 folder).